requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
tensorflow==10.0.0
pillow==9.5.0
numpy==1.21.6